+++ pkg/helm/client/client.go
@@ -292,17 +292,8 @@ func (c *HelmClient) upgrade(ctx context.Context, spec *ChartSpec) (*release.Rel
 
 	upgradedRelease, upgradeErr := upgradeAction.upgrade(ctx, spec.ReleaseName, helmChart, values)
 	if upgradeErr != nil {
-		var resultErr error
-
-		rollbackErr := c.rollbackRelease(spec)
-		if rollbackErr != nil {
-			resultErr = fmt.Errorf("release failed, rollback failed: release error: %w, rollback error: %v", upgradeErr, rollbackErr)
-		} else {
-			resultErr = fmt.Errorf("release failed, rollback succeeded: release error: %w", upgradeErr)
-		}
-
-		c.DebugLog("release upgrade failed: %s", resultErr)
-		return nil, fmt.Errorf("failed to upgrade release %q: %w", spec.ReleaseName, resultErr)
+		c.DebugLog("release upgrade failed: %s", upgradeErr)
+		return nil, fmt.Errorf("failed to upgrade release %q: %w", spec.ReleaseName, upgradeErr)
 	}
 
 	c.DebugLog("release upgraded successfully: %s/%s-%s", upgradedRelease.Name, upgradedRelease.Chart.Metadata.Name, upgradedRelease.Chart.Metadata.Version)
@@ -448,6 +439,7 @@ func mergeInstallOptions(chartSpec *ChartSpec, installOptions *action.Install) {
 	installOptions.ReleaseName = chartSpec.ReleaseName
 	installOptions.Version = chartSpec.Version
 	installOptions.Atomic = chartSpec.Atomic
+	installOptions.PostRenderer = chartSpec.PostRenderer
 }
 
 // mergeUpgradeOptions merges values of the provided chart to helm upgrade options used by the client.
@@ -459,6 +451,7 @@ func mergeUpgradeOptions(chartSpec *ChartSpec, upgradeOptions *action.Upgrade) {
 	upgradeOptions.ReuseValues = chartSpec.ReuseValues
 	upgradeOptions.Atomic = chartSpec.Atomic
 	upgradeOptions.CleanupOnFail = chartSpec.CleanupOnFail
+	upgradeOptions.PostRenderer = chartSpec.PostRenderer
 }
 
 // mergeUninstallReleaseOptions merges values of the provided chart to helm uninstall options used by the client.